mm: hugetlb controller for cgroups
In the effort of supporting cgroups v2 into Kubernetes, I stumped on the
lack of the hugetlb controller.

When the controller is enabled, it exposes four new files for each hugetlb
size on non-root cgroups:

- hugetlb.<hugepagesize>.current
- hugetlb.<hugepagesize>.max
- hugetlb.<hugepagesize>.events
- hugetlb.<hugepagesize>.events.local

The differences with the legacy hierarchy are in the file names and using
the value "max" instead of "-1" to disable a limit.

The file .limit_in_bytes is renamed to .max.

The file .usage_in_bytes is renamed to .current.

.failcnt is not provided as a single file anymore, but its value can be
read through the new flat-keyed files .events and .events.local, through
the "max" key.

Link: http://lkml.kernel.org/r/20191216193831.540953-1-gscrivan@redhat.com
Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>
Acked-by: Tejun Heo <tj@kernel.org>
Acked-by: Johannes Weiner <hannes@cmpxchg.org>
Acked-by: Mike Kravetz <mike.kravetz@oracle.com>
Cc: Michal Koutn <mkoutny@suse.com>
Cc: Li Zefan <lizefan@huawei.com>
Cc: Mina Almasry <almasrymina@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>